getAddress() can now optionally ignore SocketExceptions from NetworkInterface
Method is overloaded to keep the current behavior for all callers.
Util.getNonLoopbackAddress() however has been modified to pass ignore=true,
since this method is used for discovery and auto-configuration and if
querying a particular interface throws a SocketException, it is safe to
simply ignore that interface.

Other calls to getAddress() perhaps should ignore exceptions as well, but
this can be checked on a case-by-case basis.