Fix the Maven build - again
I found the build-helper Maven plugin allows us to add multiple source
directories to a POM, so it's possible to compile JGroups without invoking
Ant tasks.

We still need the antrun plugin to generate the schema.

Tests are compiled but some of them fail - need to configure the surefire
plugin to mimic the Ant build.